moving PhotoZShiftFactory to source
